feat : 패키지명 변경
memeber -> user
