+++ src/main/java/kr/co/studyhubinu/studyhubserver/config/jwt/JwtAuthenticationFilter.java
@@ -4,8 +4,8 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.co.studyhubinu.studyhubserver.config.auth.PrincipalDetails;
-import kr.co.studyhubinu.studyhubserver.member.dto.data.GeneralSignUpInfo;
-import kr.co.studyhubinu.studyhubserver.member.dto.request.GeneralSignUpRequest;
+import kr.co.studyhubinu.studyhubserver.user.dto.data.GeneralSignUpInfo;
+import kr.co.studyhubinu.studyhubserver.user.dto.request.GeneralSignUpRequest;
 import kr.co.studyhubinu.studyhubserver.util.CustomResponseUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.AuthenticationManager;
