+++ src/main/java/kr/co/studyhubinu/studyhubserver/config/jwt/JwtAuthorizationFilter.java
@@ -3,8 +3,8 @@ package kr.co.studyhubinu.studyhubserver.config.jwt;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import kr.co.studyhubinu.studyhubserver.config.auth.PrincipalDetails;
-import kr.co.studyhubinu.studyhubserver.member.domain.UserEntity;
-import kr.co.studyhubinu.studyhubserver.member.repository.UserRepository;
+import kr.co.studyhubinu.studyhubserver.user.domain.UserEntity;
+import kr.co.studyhubinu.studyhubserver.user.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
