+++ src/main/java/kr/co/studyhubinu/studyhubserver/user/dto/request/GeneralSignUpRequest.java
@@ -1,10 +1,10 @@
-package kr.co.studyhubinu.studyhubserver.member.dto.request;
+package kr.co.studyhubinu.studyhubserver.user.dto.request;
 
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import kr.co.studyhubinu.studyhubserver.member.dto.data.GeneralSignUpInfo;
-import kr.co.studyhubinu.studyhubserver.member.enums.GenderType;
-import kr.co.studyhubinu.studyhubserver.member.enums.GradeType;
+import kr.co.studyhubinu.studyhubserver.user.dto.data.GeneralSignUpInfo;
+import kr.co.studyhubinu.studyhubserver.user.enums.GenderType;
+import kr.co.studyhubinu.studyhubserver.user.enums.GradeType;
 import lombok.Getter;
 
 import javax.validation.constraints.Email;
