+++ src/main/java/kr/co/studyhubinu/studyhubserver/user/service/UserService.java
@@ -1,8 +1,8 @@
-package kr.co.studyhubinu.studyhubserver.member.service;
+package kr.co.studyhubinu.studyhubserver.user.service;
 
-import kr.co.studyhubinu.studyhubserver.member.domain.UserEntity;
-import kr.co.studyhubinu.studyhubserver.member.dto.data.GeneralSignUpInfo;
-import kr.co.studyhubinu.studyhubserver.member.repository.UserRepository;
+import kr.co.studyhubinu.studyhubserver.user.domain.UserEntity;
+import kr.co.studyhubinu.studyhubserver.user.dto.data.GeneralSignUpInfo;
+import kr.co.studyhubinu.studyhubserver.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
